% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.exclusion.criteria.r
\name{apply.exclusion.criteria}
\alias{apply.exclusion.criteria}
\title{apply exclusion criteria function}
\usage{
apply.exclusion.criteria(dataset, exclusion_criteria)
}
\arguments{
\item{dataset}{the dataset on which to base the report, no default}

\item{exclusion_criteria}{a list of functions defining what observations to exclude from the report, no default}
}
\description{
this function applies exclusion criteria
}
