% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.centre.dataset.r
\name{compile.centre.dataset}
\alias{compile.centre.dataset}
\title{import records function}
\usage{
compile.centre.dataset(dir_path)
}
\arguments{
\item{dir_path}{path to directory to import records from, no default}
}
\description{
this function imports records from a specific directory
}
