% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.collated.dataset.r
\name{compile.collated.dataset}
\alias{compile.collated.dataset}
\title{compile collated dataset function}
\usage{
compile.collated.dataset(dataset_name_prefix, dir_names = NULL,
  skip_dirs = NULL, save = TRUE)
}
\arguments{
\item{dataset_name_prefix}{a character string to prefix the datasets with, no default}

\item{dir_names}{a vector of directory names to import records from. default is NULL, in which case records from all directories present in path are imported}

\item{skip_dirs}{a vector of directory names to skip importing from. default is NULL, in which case no directory is skipped}

\item{save}{logical, if TRUE the datasets are saved to disk, defaults to TRUE}
}
\description{
this function puts compiles the collated dataset from records
}
