% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.data.r
\name{copy.data}
\alias{copy.data}
\title{copy data function}
\usage{
copy.data(path)
}
\arguments{
\item{path}{path to origin, no default}
}
\description{
this function copies data from origin (e.g. a server) and store here for easy and fast access
}
