% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.overview.r
\name{create.overview}
\alias{create.overview}
\title{create overview}
\usage{
create.overview(variables, data)
}
\arguments{
\item{variables}{character vector holding the names of the variables to summarise, no default}

\item{data}{the dataset, no default}
}
\description{
creates an overview, including a tableplot
}
