% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInjuryCodingReviewSet.R
\name{CreateInjuryCodingReviewSet}
\alias{CreateInjuryCodingReviewSet}
\title{Create injury coding review set}
\usage{
CreateInjuryCodingReviewSet(dataset, sample.size = 100,
  number.of.reviewers = 1, log = TRUE, save = TRUE, return.set = FALSE)
}
\arguments{
\item{dataset}{A data frame. Has the collated data from all centres. No
default.}

\item{sample.size}{A positive integer. Defines the size of the random sample
to draw from the complete set. If NULL the complete set is kept. Defaults
to 100.}

\item{number.of.reviewers}{NOT YET IMPLEMENTED. A positive integer. Defines
the number of reviewers. The set is randomly split in the same number of
parts as the number of reviewers. Defaults to 1.}

\item{log}{A logical vector of length 1. If TRUE a logfile is used to remove
observations that have already been reviewed and appended with the IDs of
observations that are included in the current set. Defaults to TRUE.}

\item{save}{A logical vector of length 1. If TRUE the review set is saved to
disk. Defaults to TRUE.}

\item{return.set}{A logical vector of length 1. If TRUE the review set is
returned to the parent environment. Defaults to FALSE.}
}
\description{
This function takes the dataset as input and creates a set of injuries and
codes for review.
}
