% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.summary.r
\name{get.summary}
\alias{get.summary}
\title{get summary function}
\usage{
get.summary(level = NULL, strata_data = NULL, t = TRUE, maxsum = NULL,
  force_na = TRUE, include_n = TRUE, include_p = FALSE, p_digits = 0,
  name = level, data)
}
\arguments{
\item{level}{the strata level to use when subsetting data based on strata_data, defaults to NULL}

\item{strata_data}{the grouping data, defaults to NULL}

\item{t}{logical, if TRUE the matrix is transposed before made into a data frame. defaults to TRUE}

\item{maxsum}{determines the maximum number of levels shown for qualitative variables, defaults to NULL}

\item{force_na}{logical, if TRUE the number of NAs is reported even if there are none, defaults to TRUE}

\item{include_n}{logical, if TRUE the number of observations is included, defaults to TRUE}

\item{include_p}{logical, if TRUE the percentage is also included, defaults to FALSE}

\item{p_digits}{integer, if include_p = TRUE this will be the number of digits in the percentage, defaults to 0}

\item{name}{(col)name of data, defaults to level}

\item{data}{the data to be summarised, no default}
}
\description{
this is a wrapper for the summary function, returns a data frame
}
