% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.timestamp.r
\name{get.timestamp}
\alias{get.timestamp}
\title{get timestamp function}
\usage{
get.timestamp(date)
}
\arguments{
\item{date}{the POSIXct vector, no default}
}
\description{
this function takes a POSIXct vector as returned by Sys.date() and returns a character vector with only numbers
}
