% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.value.labels.r
\name{get.vector}
\alias{get.vector}
\title{get vector function}
\usage{
get.vector(x)
}
\arguments{
\item{x}{the string, no default}
}
\description{
this function takes as input a single string of values or value labels and returns a vector without escape characters or whitespace
}
