% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.record.r
\name{import.record}
\alias{import.record}
\title{import record function}
\usage{
import.record(file_path)
}
\arguments{
\item{file_path}{path to file with record, no default}
}
\description{
this function imports a record from a file
}
