% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.r
\name{initiate}
\alias{initiate}
\title{initiate}
\usage{
initiate(data_path, data_copy_path, codebook_path, codebook_name)
}
\arguments{
\item{data_path}{path to data directory, no default}

\item{data_copy_path}{path to data copy directory, no default}

\item{codebook_path}{path to codebook, no default}

\item{codebook_name}{name of codebook, no default}
}
\description{
this function loads all required packages and sources all scripts
}
