% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.data.review.report.r
\name{make.data.review.report}
\alias{make.data.review.report}
\title{make data review report}
\usage{
make.data.review.report(path, report_name, report_title, data_path,
  data_copy_path, codebook_path, codebook_name, test = FALSE)
}
\arguments{
\item{path}{path to directory with records, no default}

\item{report_name}{name of the report file, no default}

\item{report_title}{title of the data review report, no default}

\item{data_path}{path to data directory, no default}

\item{data_copy_path}{path to data copy directory, no default}

\item{codebook_path}{path to codebook, no default}

\item{codebook_name}{name of codebook, no default}

\item{test}{logical, if TRUE the generated report is named test, defaults to FALSE}
}
\description{
this function compiles a data review report
}
