% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.record.r
\name{preprocess.record}
\alias{preprocess.record}
\title{preprocess record function}
\usage{
preprocess.record(text_record)
}
\arguments{
\item{text_record}{the record as text representation, no default}
}
\description{
this function preprocesses a text record to make it possible to process
}
