% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.dataset.r
\name{save.dataset}
\alias{save.dataset}
\title{save dataset function}
\usage{
save.dataset(x, dir_path)
}
\arguments{
\item{x}{the dataset to be save, no default}

\item{dir_path}{path where the dataset should be saved, no default}
}
\description{
this function saves a dataset (dataframe) to disk using write.csv
}
