% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.qualitative.r
\name{summarise.qualitative}
\alias{summarise.qualitative}
\title{summarise qualitative function}
\usage{
summarise.qualitative(variable_codebook, data, strata = list(name = NULL, data
  = NULL), verbose = FALSE)
}
\arguments{
\item{variable_codebook}{variable specific codebook, no default}

\item{data}{variable specific data, no default}

\item{verbose}{logical, if TRUE progress is printed to console, defaults to FALSE}
}
\description{
this function generates a summary of a qualitative variable
}
