% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.quantitative.r
\name{summarise.quantitative}
\alias{summarise.quantitative}
\title{summarise quantitative function}
\usage{
summarise.quantitative(variable_codebook, data, strata = list(name = NULL,
  data = NULL), verbose = FALSE)
}
\arguments{
\item{variable_codebook}{variable specific codebook, no default}

\item{data}{variable specific data, no default}

\item{strata}{list with two entries, name and data. the name entry is the name of the variable that will be used to stratify output and data is the data of that stratifying variable. both default to NULL}

\item{verbose}{logical, if TRUE progress is printed to console, defaults to FALSE}
}
\description{
this function generates a summary of a quantitative variable
}
