% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.variable.r
\name{summarise.variable}
\alias{summarise.variable}
\title{summarise variable function}
\usage{
summarise.variable(var_name, dataset, codebook, strata = NULL)
}
\arguments{
\item{var_name}{the name of the variable, no default}

\item{dataset}{the dataset, no default}

\item{codebook}{the codebook, no default}

\item{strata}{name of stratifying variable, defaults to NULL}
}
\description{
this is the master summarising function, taking as input the name of a variable, a dataset and a codebook
}
