% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.make.data.review.report.r
\name{test.make.data.review.report}
\alias{test.make.data.review.report}
\title{make weekly data review report test suite main function}
\usage{
test.make.data.review.report(path = "./test_data/server/",
  report_name = "test-weekly-data-review-report",
  report_title = "Test data review report", data_path = "./test_data/data/",
  data_copy_path = "./test_data/data/copy/", dataset_name_prefix = "test",
  codebook_path = "./test_data/codebook/",
  codebook_name = "test-codebook.csv", develop = FALSE, test = TRUE)
}
\arguments{
\item{path}{path to directory with records}

\item{report_name}{name of the report file}

\item{report_title}{title of the data review report}

\item{data_path}{path to data directory}

\item{data_copy_path}{path to data copy directory}

\item{dataset_name_prefix}{a character string to prefix the datasets with}

\item{codebook_path}{path to codebook}

\item{codebook_name}{name of codebook}

\item{develop}{logical, if TRUE the develop version of beehive is used, defaults to FALSE}

\item{test}{logical, if TRUE the generated report is named test, defaults to FALSE}
}
\description{
this function compiles the weekly data review report when testing beehive
}
