% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.report.r
\name{create.report}
\alias{create.report}
\title{create report function}
\usage{
create.report(report_title, dataset, report_name, test,
  codebook_path = .session_variables$codebook_path,
  codebook_name = .session_variables$codebook_name)
}
\arguments{
\item{report_title}{title of the data review report, no default}

\item{dataset}{the dataset on which to base the report}

\item{report_name}{the name of the report file, no default}

\item{test}{logical, if TRUE the generated report is named test, no default}

\item{codebook_path}{the path to the data codebook that governs how to interpret the data, defaults to .session_variables$codebook_path}

\item{codebook_name}{the name of the codebook, defaults to .session_variables$codebook_name}

\item{title}{report title, defaults to "TAFT data review report"}
}
\description{
this function creates the weekly data report
}
